FYP/requirements.txt
pip>=23.0
setuptools>=68.0
wheel>=0.41.0

Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0

Pillow>=10.0.0,<11.0

gunicorn
psycopg2-binary
whitenoise
dj-database-url
